requirements.txt
python-dotenv~=1.0.0
pydub~=0.25.1
tiktoken==0.4.0
openai==0.27.8
python-telegram-bot==20.3
requests~=2.31.0
tenacity==8.2.2
wolframalpha~=5.0.0
duckduckgo_search~=3.8.3
spotipy~=2.23.0
pytube~=15.0.0
gtts~=2.3.2
whois~=0.9.27
waitress
flask[async]
gunicorn==20.1.0
aiofiles==23.2.1
langchain==0.0.315
psycopg2-binary==2.9.9
pgvector==0.2.3
tiktoken==0.4.0
